+++ config/credentials.sample.yml
@@ -8,6 +8,7 @@ default: &default
 
   contentful:
     space_id: XXX
+    webhook_password: CreateYourOwnStrongPassword!
     delivery_access_token: XXX
     preview_access_token: XXX
